services/technical-indicators/pyproject.toml
[project]
name = "technical-indicators"
version = "0.1.0"
description = "Technical indicators calculation service using pandas-ta"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru",
    "quixstreams",
    "pydantic-settings",
    "pandas",
    "pandas-ta",
    "pyyaml",
]

[project.scripts]
technical-indicators = "technical_indicators.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-mock>=3.14",
]

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
